requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.2
cohere==4.34
colorama==0.4.6
dataclasses-json==0.6.2
distlib==0.3.7
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
fastavro==1.8.2
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
huggingface-hub==0.19.3
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.336
langsmith==0.0.64
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.4
packaging==23.2
platformdirs==3.11.0
pydantic==2.5.1
pydantic_core==2.14.3
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
tenacity==8.2.3
tokenizers==0.15.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.6
virtualenvwrapper-win==1.2.7
yarl==1.9.2
zipp==3.17.0
